AdminAPI test fixes:
Added check_sandbox_in_metadata() in the available test functions to
wait for an instance which was recently added to a cluster to be
present in its own metedata schema: there may be some replication
delay since the instance is added to the metadata on the primary.

Included the usage of check_sandbox_in_metadata() for the
dba_cluster_add_instance and super_read_only_handling tests.
Those tests were failing due to the assumption that if an instance
has the "ONLINE" state it already has all the transactions applied
and therefore has the most recent version of the metadata schema.
This may not be always true due to the slowness of the system
running the tests.

Added connection attempt in try_restart_sandbox() to verify if the
instance is up-and-running and accepting connection before returning.

Replaced the use of killSandboxInstance() with stopSandboxInstance()
in several tests to make sure the instance is completely dead before
restarting it to make sure the restart does not fail. Made use of
try_restart_sandbox too instead of startSandboxInstance().