AdminAPi test fixes:
Added check_sandbox_has_metadata() in the available test functions to
wait for an instance which was recently added to a cluster to have
the metedata schema before proceeding to any operation which requires
the metadata: there may be some replication delay.

Included the usage check_sandbox_has_metatadata() in
check_sandbox_in_metadata(): in some cases the metadata schema wasn't
present yet in the instance so the checks in the 'instances' table
failed.

Replaced the use of killSandboxInstance() with stopSandboxInstance()
in super_read_only_handling test to make sure the instance is
completely dead before restarting it to make sure the restart does not
fail. Made use of try_restart_sandbox too instead of startSandboxInstance().