BUG#26001653: CLUSTER.DISSOLVE() FAILS IF ONE OR MORE INSTANCES HAVE THE (MISSING) STATUS
The Cluster.dissolve() function was trying to stop group-replication on
all the instances registered in the Metadata which leads to connection
errors if any of those instances have the "(MISSING)" status.
This patch fixes the issue by making sure that group-replication is
stopped only on the instances which have the "ONLINE" status.

Added new tests to verify the changes above and also to fully test the
dissolve() operation in single-primary and multi-primary scenarios.