BUG#26636911 : AdminAPI commands throw error that does not support SSL for Win
The AdminAPI commands throw an error when used on servers with SSL support
on Windows because the Python version used for that platform is the one
bundled with the shell which does not have SSL support.

In non-Windows systems this does not happen because the Python available in
the system is used (not the one bundled with the shell). For those platforms
there is no issue as long as the system Python supports SSL. This issue
was manifested because secure connection are performed by default and to
perform them connector/python requires a Python version with SSL support to
be used.

This patch replaces the use of connector/python with the use mysql-shell
sessions to establish connections to the server for the internal
mysqlprovision tool, always using the Python bundled with the shell
independently of the platform, but no longer requiring Python to have SSL
support.