BUG#26979375 : ADMIN API COMMANDS FAILING WHEN THE USER CONNECTED TO ROOT@%
This patch fixes an issue with the account validation being done on
the check_instance_configuration method that did not work correctly
unless the session account existed.  With this patch, that validation
is now done using the account that was authenticated by the server,
i.e., the account returned by the SELECT CURRENT_USER() SQL query.
Furthermore, it now also makes sure that the correct account is used
by the MP backend on the check command.

New tests were added to verify the fix.