Fixed AdminAPI test failures in windows:
In windows platforms some tests were failing due of an error in
dba.deleteSandboxInstance() calls as the directories were still being
accessed. As dba.stopSandboxInstance() returns without full guarantee
that the mysqld process if stopped leading to a possible failure in
deleting the sandbox dir aftewards.
The fix was to include a retry method with a 60 seconds timeout for
the deleteSandboxInstance calls in the cleanup_sandbox() function.

Reverted commit 6b2190ebbce76edc354e06169fcee8f488aa5006 and included
patch to disable by default the Shell_py_dba_tests