WL#11162 : Add options to set GR address, seeds and name for a cluster
Fix a regression issue in the doxygen documentation generation. Warnings were
issued when generating the documentation due to the use of <host> and <port>
in the help for the new options added by this feature. The use of '<' and '>'
was removed to solve the problem.

The existing tests were updated accordingly.