Fixed Windows test failures:
Command_line_test.bug24905066 is failing on windows because the Shell is
not recognizing the named-pipe identifier (.) as equivalent to
'localhost'. A possible solution would be to use 'localhost' but this
actually needs to be fixed in the shell.

Shell_js_dba_tests.configure_local_instance is failing on windows
because of a wrong log validation which is expecting the string
"mysqlprovisioning" instead of "mysqlprovisioning.zip"

Shell_js_dba_tests.dba_cluster_add_instance is failing on windows
because of a wrong path splitting logic.