From MySQL Server 8.0.4 we use system OpenSSL on most systems. This is already handled in the Shell sources in that it will link to the client library (that not lacks bundled SSL objects) and use the output from "mysql_config" to add shared library dependencies.
But the server packages will not use system OpenSSL on Windows and macOS, as there is
no default install of OpenSSL on those operating systems. Instead shared libraries are
bundled. In the case of Windows in "bin/" enabling the binary to find the DLLs, in the
case of macOS in "lib/" and the executable using a relative RPATH to find them.

This patch adds the same feature to MySQL Shell for Windows and macOS.