BUG#26796118 : instance rejoins GR group after removeInstance() and restart
After removing an instance from a cluster, using the removeInstance()
function, the instances silently rejoins the Group Replication (GR) group
after it is restarted. This happens because the GR plugin automatically start
GR when the server is restarted (group_replication_start_on_boot=ON by
default).

This patch fixes this issue by setting group_replication_start_on_boot to
OFF, using SET PERSIT for server version >= 8.0.2. For server versions
without the SET PERSIST feature a warning is issued to notify the users
that the must persist this setting on the configuration file manually to
avoid this issue.

In addition, this patch also fixes a few issues in the test suite:
- improve validation of the TMPDIR environment variable, required for unit
  tests;
- fix sandbox self deployment features, to allow tests to be execute
  individually;
- improve testutil.destroySandbox() to provide an option not to stop the
  server, since it might be already stopped (avoiding errors trying to kill
  it);

Appropriate unit tests were added for regression.