WL#11344: InnoDB Cluster configuration automatically persisted
For server version prior to 8.0.0, it was not possible to persist
change to server configuration remotely, i.e., without local access
the configuration file. With the introduction of the SET PERSIST
statement it is now possible to remotely persist changes to server
configurations.

This patch adds support to automatically persisting Group replication
and other required server settings remotely for cluster operations
(create a cluster, configure, add, or remove an instance) using the
AdminAPI.  This is only supported for server version >= 8.0.4.

New tests were added to exercise the new features implemented by this
WL and some existing tests had to be modified.