BUG#27193423 : DBA.DEPLOYSANDBOXINSTANCE: (RUNTIMEERROR)
Added a call to PyErr_Print() to print the traceback to sys.stderr
whenever the import for the module 'runpy' fails in the Shell Python
context handler (CPython). Such call was added for debugging purposes
for this bug, however, after it the bug could not be reproduced anymor.
The reason is probably related with an internal CPython resource leak.