MP: Preventing server monitoring process on sandbox operations
Starting 8.0.4 the windows server starts a second mysqld process
that monitors the real server (related to the RESTART operation).
Because of that, the sandbox deployment was getting stuck.
This fix is to define a MYSQL_PARENT_PID environment variable
that will prevent the monitoring process from being launched
and will let hte server working as before.

UT: Added retry logic for the boilerplate sandbox operation
which was failing in windows.