BUG#27089930 : forceQuorumUsingPartitionOf does not handle non root user
The AdminAPI function forceQuorumUsingPartitionOf() does not wort with a
non 'root' user, even if that user has all the required privileges. This
happens because internally this operation is using the 'root' user to connect
to other instances instead of the specified user.

It is assumed that the same user (cluster administration user) with the same
authentication credentials and privileges is available on all the instances,
but that user should not be forced to be the 'root'. The user specified
with the clusterAdmin option of the dba.configureLocalInstance() function
is expected to be used as cluster administration user on all instance to
execute any AdminAPI operation.

This patch fixes that issue by using the specified user to connect to all
instances, instead of the previously hard-coded 'root' user.

Test was added for regression.