BUG#26986141 : removeInstance() not accepting the second parameter
In interactive mode the <cluster>.removeInstance() AdminAPI function is
not accepting the second parameter with the options dictionary, for example
the use of the 'force' option, issuing an error.

This patch fixes this issue by correcting the connection parsing. It also
improves the code to ignore any connection error that could occur when trying
to check the server version if the force option is used (since in that case
it is expected that the instance is not available).

A proper interactive test was added for regression (non-interactive test was
already available).