BUG#27061615: adopting cluster from existent GR group in wizard mode has misleading output
Whenever a cluster is adopted from an existent GR group by using the option 'adoptFromGR' the output is the same as if the option was not specified.
This is misleading as it suggests that additional action is needed when that's not true.

This patch detects if 'adoptFromGR' option was used and provides a suitable output to the user.

Added more test cases to Shell_js_dba_tests.adopt_from_gr_interactive.
Fixed tests affected by the implementation change.
Trace files recreated to match changes in unit tests.