BUG#27095984 : addInstance() fails in Debian platforms when using hostname
When creating a cluster in a Debian platform (e.g., Ubuntu) using the default
system configurations the <cluster>.addInstance() command fails with an error
if the hostname is used in the connection parameter. This happens because in
Debian platforms the hostname is resolved to the IP address 127.0.1.1 by
default, but this IP address is not supported by the Group Replication
Communication Service (GCS).

This patch add a validation to the dba.createCluster() and
<cluster>.addInstance() functions to verify if the connection hostname
resolves to 127.0.1.1 (not supported by GCS) and issue an error in that case.