BUG#27146290: AdminAPI online helper generic connection info should be at the end
The online helper for all AdminAPI commands which require a connection displays
detailed information on connection data, which clutters up the details of each
command.

This patch removes the detailed information on connection data and provides a
hint on where to look for details.

Unit tests updated to match changes in the implementation.