BUG#26315635: unable to execute dba.dropMetadataSchema() on stand-alone instance with metadata
The AdminAPI does not check if a stand-alone server instance has metadata schema,
requiring the instance to belong to InnoDB cluster in order to execute the
dba.dropMetadataSchema() function.

This patch introduces a new internal instance state to identify a stand-alone instance with
metadata schema and allows to execute the dba.dropMetadataSchema() function on such
instance.

Added unit tests to verify if it's possible to drop metadata schema on
a standalone instance which has the metadata.
Traces have been recreated due to implementation changes.