BUG#27076753: user missing privileges error message lacking detail
Two AdminAPI functions, configureLocalInstance() and checkInstanceConfiguration(),
check if the account being used has enough privileges to administer an InnoDB
cluster, but the error message shown when the account doesn't have all of them
lacks information on which privileges are missing.

The method which verifies the account was modified in order to provide detailed
information on missing privileges, and to identify the schema they apply to.

Unit tests updated to match changed output.
Added unit tests for user with no privileges.
Traces have been recreated due to implementation changes.