BUG#27437389: rejoining an instance to the cluster in interactive mode has misleading output
Rejoining an instance in the interactive mode produces two similar lines of
output, containing duplicated information.

The second line is actually the return value from the rejoinInstance()
function, since this function should not return any values, it was removed.
The output message was modified to contain host:port of rejoined instance,
the username was removed.

Unit tests have been adapted to the new output.