BUG#27181177 : dba commands fail if user name has non-standard characters
The AdminAPI commands to manage sandboxes fail with a RuntimeError when using
sandbox directory with non-ASCII characters. This occurs if the user name
has non-ASCII characters because the default sandbox directory is a
subdirectory of the $HOME or %userprofile%, which is usually named according
to the used user name.

This patch fixes this issue by adding proper Unicode support to the
internal provision tool (in Python) used by the AdminAPI.

New unit tests were added for regression.