BUG#26962715 : addInstance() does not check for unique server_uuid
The <cluster>.addInstance() function does not check the validity of the
server UUID being added to the cluster. The server UUID of all members of
the cluster must be unique (replication requirement) and an proper error
should be issued to prevent adding instance with the same UUID of an
existing cluster member.

This patch fixes this issue by verifying if the server_uuid of the instance
to add is different from the server_uuid of all current members of the
cluster, issuing an error if an instance with the same server_uuid is
found.

New unit tests were added for regression.