BUG#25042407: dba.createCluster() cannot handle ::1 host of seed node from shell login
The MySQL shell is able to create a session to an IPv6 address, but it fails
to create a cluster using such connection, reporting an URI-related error.

The bug report allowed to identify an issue related to encoding of the URI
containing an IPv6 address, which was fixed by this patch. The core issue,
IPv6 support for AdminAPI, cannot be fixed, as group replication requires IPv4
network: https://dev.mysql.com/doc/en/group-replication-requirements.html.
The implementation of cluster-related AdminAPI was modified in order to check
if operations are executed using IPv4 session and IPv4 connection data,
throwing an exception if these conditions are not met.

Updated the list of exceptions of AdminAPI methods.
Updated unit tests, connection data requires a valid hostname.
Updated unit tests of Connection_options to check if IPv6 is handled properly.
Added unit tests for AdminApi via IPv6 connection.
Traces have been recreated due to unit test changes.
Recorded traces for new tests.