BUG#24761416: cleanup after a failed addInstance() call not proper
Adding two different cluster instances with the same label correctly results in
an error, however the instance with the duplicated label is left with partially
initialized configuration.

Added a check to determine if the requested label is unique in a cluster,
duplicated label results in an exception being thrown.

Added unit test to check validation of duplicated label.
Recorded traces for new unit test.
Traces for affected tests have been recreated.