BUG#27572618 : MySQL Shell creates a root user missing privilege on SYS
The removeInstance() command was issuing an error reporting missing
privileges when using a (admin) user created by dba.configureLocalInstance().
In more detail, the error indicated that the execute command was denied to
the user for routine 'sys.version_major'.

This issue was already fixed by the patch for BUG#26986141, which removed the
use of the sys.version_major(), sys.version_minor(), and sys.version_patch()
to obtain the server version information, no longer requiring EXECUTE
privileges on the "sys" schema.

This patch adds a test for regression.