BUG#26836230 : check and configure not showing server_id issues
The dba.checkInstanceConfiguration() and dba.configureLocalInstance() commands
do not properly display the error related with server_id misconfiguration.
An error is reported but there is no information about the problem being
related with the server id variable (errors is empty).

This patch fixes this issue by properly displaying the server_id in the list
of variables not meeting the requirements. The server_id change requires a
restart of the server to be properly applied.

In addition, this patch also removes the X plugin load from the my.cnf for
sandbox version >= 8.0.5, since it is now automatically available by default
for those versions.

A test case was added for regression.