BUG#27545850: GR options cannot be set while start/stop GR plugin operation in progress
The 8.0.5 version of the server introduced a new check, it is not possible to
set the GR variables while GR plugin is starting/stopping.

This patch introduces a new precondition check for all AdminAPI operations,
requiring that GR plugin is not currently starting, issuing an error if this
condition is not fulfilled. It's not possible to check if GR plugin is
stopping, hence such condition was not added.

Recreated traces due to implementation changes.