BUG#26754410 : configure instance adds MyISAM and CSV to disabled_storage_engines
The dba.configureLocalInstance() command added MyISAM and CSV to the
configuration without any notification. InnoDB Cluster, more precisely Group
Replication requires the InnoDB storage engine, not supporting other engines.
Therefore, in order to ensure that the correct engine is used the
dba.configureLocalInstance() sets disabled_storage_engines to disable any
other engines. Nevertheless, this change cannot be performed silently since
disabled_storage_engines is not a dynamic variable and a server restart is
required for the option changes to be applied, properly notifying users that
the option was changed and to take the required action (restart the server).

This patch fixes this issue by properly handling disabled_storage_engines like
other non dynamic required server configurations.

A test case was added for regression.