BUG#25061891, BUG#25664766: createCluster() with adoptFromGR:true on an instance without GR should result in error
Both bug reports indicate that creating a cluster with option adoptFromGR:true
on an instance without an unmanaged group replication should be detected and
reported as erroneous.

Added a validation which checks if the adoptFromGR:true option is used to create
a cluster on an instance with unmanaged group replication, an error is reported
if this validation fails.

Added unit tests for adoptFromGR option without replication group.
Recorded traces for affected tests (5.7.21).
Recorded traces for affected tests (8.0.11).