BUG#27329079 : create cluster not possible with innodb_page_size=4k
The dba.createCluster() function fails if used on an instance with
innodb_page_size=4k, because the instance_name column of the instances table
in the InnoDB Cluster metadata uses a VARCHAR of size 256 (to large for the
innodb_page_size=4k). However, the size of this column cannot be changed
because it is used to hold hostnames of the cluster members which can have a
lenght up to 255 bytes. This limitation should be validated and documented.

This patch adds a innodb_page_size validation to the AdminAPI operation to
check and configure an instance, as well as for creating a cluster, reporting
an error if the instance innodb_page_size is not supported.

Test was added for regression.