WL#11862 AdminAPI: Refactor removeInstance()
Patch with the implementation of removeInstance() fully in C++. In particular,
with the conversion of the mysqlprovision command leave-replicaset to C++,
and refactoring of the <cluster>.removeInstance() command to use the new class
structure. In general, the implementation maintains the current behavior of
the removeInstance() operation. However, the output of the command was
improved, an 'interactive' option was added, and a new shell option
'dba.gtidWaitTimeout' was added. In particular, now in interactive mode users
will be asked to continue with the operation or not if the target instance is
not reachable.

In addition, the following fixes are also included in this patch:
- replace any mention to MySQL server version 8.0.5 (which does not exists)
  by version 8.0.11 (which is the correct server version);
- replace WAIT_UNTIL_SQL_THREAD_AFTER_GTIDS() by WAIT_FOR_EXECUTED_GTID_SET()
  in Testutils::wait_member_transactions(), since the later function wait for
  all transaction to be applied independently of the channel;
- fix BUG#27817894 by waiting for transactions to be applied on the instance
  to remove before actually removing it from the cluster;

Tests added for regression to exercise the new changes.