BUG#25966235 lack of the needed user permission, ambiguous error message
Creating a cluster using an account which is missing the global GRANT option
fails with ambiguous error message, even though checkInstanceConfiguration()
did not return any errors.

Added verification to check if the account which is going to be used to
administer the cluster has the global grant option.

Added unit tests which reproduce the reported bug.
Recorded traces for affected tests.