BUG#26754410 : DBA.CONFIGURELOCALINSTANCE EXPLICITS ADDS MYISAM, CSV TO DISABLEDSTORAGENGINES
It has been defined in the past that one of the requirements for InnoDB
Cluster is that the following storage engines must be disabled:
MyISAM,BLACKHOLE,FEDERATED,CSV,ARCHIVE, thus
dba.checkInstanceConfiguration() and configureLocalInstance() report
invalid configurations if those are not disabled in the targed instance.
In top of that, dba.configureLocalInstance() does explicitly change the
instance configuration in order to include those as disabled storage
engines. Also, the sandbox deployment takes this in consideration and
the configuration files of the sandboxes disable those storage engines.
However, the requirement is too strict and is actually a recommendation
and not a requirement so the decision is to remove it.

This patch removes the requirement from the checks and sandbox
deployment.

Updated tests and traces to meet the changes.