BUG#27746532: APPLICATION ICON MISSING
A resource file is required in windows to assign the icon to the application,
it also defines the version information that will be shown while looking at the
file properties.

A recfatoring caused this resource file to NOT be associated to the main
application, instead it was being associated to the new frontend library,
causing the icon and the version information to not be associated correctly.

Fix consist in re-associating the resource file correctly, to the main
application.