BUG#27758041 : SHELL PICKS RANDOM PORTS IF GR LOCAL ADDRESS PORT IS BUSY
When creating a cluster or adding an instance, if not localAddress option
is specified, the port used for group_replication_local_address is
automatically assigned with the value: mysql_port * 10 + 1. However,
if the the resulting port determined by the previous rule is already in
use then a random port is generated and used.

This patch changes the previously described behavior by issuing an error
instead of generating a random port when it resulting port from
mysql_port * 10 + 1 is already in use.

Test was added for regression.