BUG#27765484 : BIND_ADDRESS IN MY.CNF NOT FIXED BY CHECKINSTANCECONFIGURATION/CONFIGUREINSTANCE
It has been defined in the past that one of the requirements for InnoDB
Cluster is that 'bind_address' must be unset or set to the default
value, thus
dba.checkInstanceConfiguration() reports an invalid instance
configuration if
'bind_address' is set to a value not the default one on the targed
instance.
In top of that, dba.createCluster(), since it validates the instance
configuration before proceeding with the operation, fails with an error
indicating the instance configuration is invalid.

However, such requirement is too strict and is actually a recommendation
and not a requirement from GR so the decision is to remove it.

This patch removes the requirement from the checks above mentioned.

Added tests to meet the changes.