BUG#27809310: SHELL.GETSESSION() CAUSES CRASH IF NO SESSION IS ESTABLISHED
A crash was generated when the function was called and no session
was established. The problem is that when a null object is assigned
to a Value, it still indicates it is an object and when attempting
to translate the value to JS/PY it crashes.

The fix consist on adding validation to the Value constructor to
make it set the type to NULL if the assigned object is Null.

(cherry picked from commit 8e1858b65f71e42ee8f4fe342a6240fbcb69d159)