BUG#24702489: inconsistent reconnect behaviour
The MySQL Shell is able to automatically reconnect global session when running
in the interactive mode, but AdminAPI methods lack this feature.

The Dba-related methods which utilize the global session object have been
improved in order to detect an interrupted session and trigger the
reconnection mechanism. The Cluster object uses its own internal session
instance, which does not support the reconnection. If connection is lost, user
needs to recreate the Cluster object.

Change-Id: Ibc6da0be6565559a1d0491ae05d17c8196b99633