WL#11892 Upgrade checker: add JSON output, externalize checks descriptions
This WL enhances Upgrade Checker tool with the ability to produce machine friendly JSON output.
The function interface had to be amended to incorporate new parameters:

Integer checkForServerUpgrade(ConnectionData connectionData, Dictionary options)

Both arguments are optional, connectionData must be given if Shell session is not connected
to server. The second argument - a dictionary will be used to handle all the parameters
that are not related to connection options. As part of this WL two parameters will be handled:
- password - password for connection,
- outputFormat - with possible values TEXT (default) and JSON.

This WL also enables storing textual content of the tool in external text file
"upgrade_checker.msg" to speed up development and give ability to provide users
with more comprehensive explanation of changes.

BUG#27927641 - turn off zerofill check for 8.0 has also been fixed in this work.