BUG#27062122 MEMBERSSLMODE NOT NEEDED FOR ADDINSTANCE() AND REJOININSTANCE()
The memberSslMode option is not needed for the <cluster>.addInstance() and
<cluster>.rejoinInstance() operations because the SSL mode must be the same
for all cluster members and the correct value is obtained automatically for
those operations when not specified.

This patch deprecates the use of the memberSslMode option for
<cluster>.addInstance() and <cluster>.rejoinInstance() operations, issuing a
warning when the option is used for those operations and updating the help to
indicate that the option is deprecated and will be removed in a future
release.

Tests were updated for regression.