BUG#28064621 GROUP_REPLICATION_FORCE_MEMBERS MUST BE UNSET AFTER QUORUM RESTORED
The <cluster>.forceQuorumUsingPartitionOf() operation sets the variable
group_replication_force_members on the target instance to force a new group
membership (restore the quorum), but it does not clean the value of the
variable at the end. Consequently, if Group Replication later needs to be
restarted on the target instance it will fail because the "force members"
variable is still set (not empty).

This patch fixes this issue by clearing (i.e., set to an empty string) the
group_replication_force_members variable at the end of the
<cluster>.forceQuorumUsingPartitionOf() operation.

A test was added for regression.