BUG#25926603 WRONG GROUP REPLICATION TERMINOLOGY USED THROUGHOUT THE ADMINAPI
This patch replaces the use of the multi-master and single-master
terms throughout the code with multi-primary and single-primary
respectively. It also deprecates the multiMaster option in the
dba.createCluster command, introducing a new multiPrimary option.
Existing tests were updated accordingly and some new tests were also
added to show the deprecation of the old option.

Change-Id: Ic7b1f73a87e8c2d23636756ad75ae0cc5ff6cc61