BUG#27148943 GETCLUSTER() SHOULD WARN WHEN CONNECTED TO MEMBER WITH NO QUORUM
The dba.getCluster() operation does not display any warning to the user when
connected to a member in a cluster with no quorum. Users should be warned
about this situation since a cluster with no quorum cannot process write
transactions.

This patch fixes this problem by issuing a warning to the output (not only to
the log) when the cluster retrieved by the dba.getCluster() operation does not
have quorum. In addition, the previous warning message that was only sent to
the log was improved.

Test was added for regression.