BUG#27508698 forceQuorumUsingPartitionOf() allowed on cluster with quorum
The forceQuorumUsingPartitionOf() operation should not be allowed to be
executed on a cluster with quorum. As a result, currently a MySQL Error
indicating that the variable 'group_replication_force_members' cannot be set
is issued, which is not user friendly.

This patch fixes this issue by adding a validation to verify if there is
quorum on the cluster (based on the perspective of the used target instance),
issuing an error in that case to stop the execution of the operation. In
addition, the internal session handling (closing sessions) was improved.

Test was added for regression.