WL#10587 BUG#26140094 BUG#28165891
WL#10587: AdminAPI: ipWhitelist filtering should apply to InnoDB Cluster
internal accounts

The Shell (AdminAPI) supports defining which hosts are allowed to be
part of a cluster. This allows increasing the security of the cluster by
limiting the network accesses to engage group communication with the
cluster with the intent of joining it. This can be defined when creating
a cluster or when adding/rejoining instances to a cluster, by using the
option 'ipWhitelist'.

However, even though the AdminAPI supports this security wise feature,
whenever the 'ipWhitelist' option is used the internal replication-user
created is not honoring the defined 'ipWhitelist' filters. Currently,
the replication-user is created with the '%' wildcard as host value thus
not respecting the security enforcement set by 'ipWhitelist'. On top of
that, it is a common practice (by companies) that the password policy
does not allow the '%' wildcard to be used in the hostname.

This patch ensures that the internal InnoDB cluster accounts use the
same filtering logic as defined by the 'ipWhitelist', i.e. the created
user accounts can connect exclusively from the allowed hosts specified
in the 'ipWhitelist' option. The accounts are created using equivalent
netmask values for all the 'ipWhitelist' values represented with the
CIDR notation.

On top of that, <Cluster>.rejoinInstance() has been improved to,
whenever 'ipWhitelist' is used, remove the previously existing internal
account from all the cluster members and create a new one to match the
'ipWhitelist' filter.

No visible changes to the user have been introduced.
New tests added to cover the changes and new functionalities added.

Bugs fixed:

BUG#26140094 - NEED OPTIONS TO SPECIFY A HOST OR IP FOR AUTO-GENERATED
USER
BUG#28165891 - ENHANCEMENT ERROR MESSAGE WHEN GIVING WRONG IPWHITELIST
ARGUMENT

Change-Id: I6463b607f8993400961ee107b4c0938f75a63b67