BUG#27508627 rebootClusterFromCompleteOutage should not point to force quorum
The rebootClusterFromCompleteOutage() operation suggest the use of the
forceQuorumUsingPartitionOf() operation when trying to execute the operation
on an instance that already belongs to a InnoDB Cluster and that is reachable,
independently if there is quorum or not. The indication to use
forceQuorumUsingPartitionOf() in the error message is misleading and should
be removed, especially if there is quorum.

This patch fixes this issue by introducing an additional quorum verification
to the existing validation and only adding the indication to use
forceQuorumUsingPartitionOf() in the error message if quorum was in fact lost
according to the checked instance.

Test was added for regression.