BUG#27765769 SHELL 8.0.X SHOULD BLOCK ADMINAPI OPERATIONS FOR 9.0.X SERVERS
MySQL Shell 8.0 will not support AdminAPI operations on MySQL Server 9.0
but the operations are not currently blocked which will result in
inconsistent behaviours and errors.

This patch introduces a validation mechanism to block all AdminAPI
operations if the target instance has a version > 8.0 or < 5.7.
In such scenario, the operations terminate with an exception indicating
that the target instance version is not supported.

Added unit-tests for the new behaviour.

Change-Id: I91337a7157b976111e646be2aace1eda06d6874f