Bug#28134957 - SHELL CANNOT CONNECT TO 8.0.11 SERVER ON X PORT
Workaround for bug#28134957, caused by a bug in libmysqlxclient.
Authentication sequence implemented in libmysqlxclient as of 8.0.12
aborts the process as soon as a fatal error is seen, which will work
if the server is 8.0.12, but fails if it's 8.0.11.

In order to make the authentication code work with any server version,
the fatal error is ignored so that other authentication methods can
be attempted.

Workaround should be reverted once bug#28135006 is closed.

Change-Id: Ifdbd510f6f54305c7a6fa966e1e2e342e48199e5