BUG#27344040 rebootClusterFromCompleteOutage should not create new user
The rebootClusterFromCompleteOutage() operation creates a new recovery
(replication) user on the target instances. The creation of a new user is
unnecessary since the operation is used to reboot a cluster with already
configured InnoDB Cluster members (already with the required recovery users).
This issue will lead to the existence of an increasing number of necessary
users.

This patch fix this issue by skipping the creation of recovery (replication)
users for the rebootClusterFromCompleteOutage() operation.

Test was added for regression.