BUG#27539702 ILLEGAL CHARACTER '%' WHILE USING SHELL.CONNECT COMMAND
The '%' character separating IPv6 address from zone ID was not handled
properly.

This patch introduces support for IPv6 addresses with zone ID. Zone ID should
be a non-empty string containing unreserved URI characters or URL encoded
sequences.

If IPv6 address with zone ID is used to specify a host, it can be used
directly:
- mysqlsh --user=user --host=fe80::850a:5a7c:6ab7:aec4%enp0s3 --port=33060

If such address is used in URI, the '%' character needs to be URL encoded:
- mysqlsh --uri=user@[fe80::850a:5a7c:6ab7:aec4%25enp0s3]:33060
- shell.connect("user@[fe80::850a:5a7c:6ab7:aec4%25enp0s3]:33060")

Change-Id: I5ef96aec000dd2d4dc4932c97e56362a4f4c456a