WL#11889 AdminAPI: Refactor dissolve()
Patch with the implementation of <cluster>.dissolve() operation fully in C++.
In particular, removing the use of the mysqlprovision command
leave-replicaset and refactoring of the <cluster>.dissolve() command to use
the new class structure. In general, the implementation maintains the current
behavior of the dissolve() operation, except for the usage of the 'force'
option which is now used to decide to continue with the operation in case
any of the cluster members has some issue (e.g., not reachable). Apart from
this main behavior change, the following changes were also made: a new
'interactive' option was added; new prompts were added in interactive mode
for the user to confirm the execution of the operation or decide to continue
in case an issue is detected and no force option is used; the output of the
command was improved.

In addition, the following fixes are also included in this patch:
- remove an unused (leftover) disable operation from the replicaset module;
- fix BUG#27837231 by no longer requiring to always use the force option
  even if there is no issue with any cluster member;
- fix BUG#27833605 by waiting for transactions to be applied on each instance
  before actually removing them from the cluster;
- fix regression (for BUG#25966057) on AdminAPI code when enabling ANSI_QUOTES
  SQL mode;
- fix issue in the regression test for BUG#26836230 when running with 5.7
  servers;

Tests added for regression to exercise the new changes.

Change-Id: Ibe92efe40ab835f134208baf1335a1bbd944f26f