BUG#27819540 FORTIFY: COMMAND INJECTION ON PROCESS_LAUNCHER
To execute mysqlprovision operations the Process_launcher uses a
global variable that is initialized with argv[0].

Fortify reports that as a problem by the fact that there's now
guarantee that argv[0] contains the path to the current executable.

The problem has been fixed by splitting get_binary_folder() function
into two: get_binary_path() which returns the path to the binary
and get_binary_folder() itself, which only trims the program name from
the returned path.

The new function get_binary_path() is now used to initialize the global
variable.

Change-Id: I44615d603d4b70c23f0c0daab9531a7cde929589