BUG#27894642 MYSQL SHELL IS CLOSED IF THE USER USE THE COMMAND ^C IN WINDOWS
The Ctrl-C command on Windows was not handled properly when reading user input,
resulting in shell application being closed.

This patch fixes the Ctrl-C handling on Windows and enables unit tests for
keyboard interrupts on that platform.

Note: The issue when Ctrl-C needs to be repeated in order to cancel a command
is also observable on non-Windows platforms, though the time frame when Ctrl-C
can be ignored is a lot smaller. Shell is not always able to immediately
interrupt execution, in this particular case it's invoking another process to
deploy a sandbox and until that process is up and running interruption is
impossible.

Change-Id: I0ad7eb0c2205f2ad51a53bd0c46fd7cdf9a19b31