BUG#27068352 WHEN EDITING LONG COMMANDS FROM HISTORY MYSQL SHELL OVERLAP THE COMMAND
On Windows, editing a command at the right edge of the console window
makes it move one line up, overwriting the output of previous commands.

Issue was caused by wrong positioning of the cursor when command reached
right edge of the console window and fixed by moving the cursor one line
down.

Change-Id: I84421a9be0f6a9e78df6e2854398fc9a75f0e212